render: modularize
This makes the render components reusable between the different modes.
At the moment `object-as-message`, and `error` are using its own
rendered - and with browser mode potentially being added, this might
become even more rough. This patch aims to remove as much code as
possible, catering to simplify internals so that it not only becomes
eaiser to maintain, but it also becomes viable to include in the
browser.

In order to lose weight some feature have been cut. Mainly: passing in
custom colors caused a lot of complexity, but also log level related
options were rough.

---

Breaking changes
================
- messages now require a `level` property to be considered renderable,
  rather than if a key exists the message occurs as log level `info`
- passing in `name` into options will no longer cause the `name` field
  to be ignored on messages
- the option to override colors has been removed
- the option to set a default name has been removed
- log level cutoff can no longer be set

Bug fixes
=========
- object mode wasn't wrapping time in `chalk.dim()`
- error mode wasn't wrapping time in `chalk.dim()`